+++ libraries/javalib/gnu/xml/libxmlj/dom/GnomeCharacterData.java
@@ -39,53 +39,59 @@ extends GnomeNode
 implements CharacterData
 {
 
-  GnomeCharacterData(int id)
+  GnomeCharacterData (long id)
     {
-      super(id);
+      super (id);
     }
 
-  public String getData() throws DOMException
+  public String getData ()
+    throws DOMException
     {
-      return getNodeValue();
+      return getNodeValue ();
     }
 
-  public void setData(String data) throws DOMException
+  public void setData (String data)
+    throws DOMException
     {
-      setNodeValue(data);
+      setNodeValue (data);
     }
 
-  public int getLength()
+  public int getLength ()
     {
-      return getData().length();
+      return getData ().length ();
     }
 
-  public String substringData(int offset, int count) throws DOMException
+  public String substringData (int offset, int count)
+    throws DOMException
     {
-      return getData().substring(offset, offset + count);
+      return getData ().substring (offset, offset + count);
     }
 
-  public void appendData(String arg) throws DOMException
+  public void appendData (String arg)
+    throws DOMException
     {
-      setData(getData() + arg);
+      setData (getData () + arg);
     }
 
-  public void insertData(int offset, String arg) throws DOMException
+  public void insertData (int offset, String arg)
+    throws DOMException
     {
-      String data = getData();
-      setData(data.substring(0, offset) + arg + data.substring(offset));
+      String data = getData ();
+      setData (data.substring (0, offset) + arg + data.substring (offset));
     }
 
-  public void deleteData(int offset, int count) throws DOMException
+  public void deleteData (int offset, int count)
+    throws DOMException
     {
-      String data = getData();
-      setData(data.substring(0, offset) + data.substring(offset + count));
+      String data = getData ();
+      setData (data.substring (0, offset) + data.substring (offset + count));
     }
 
-  public void replaceData(int offset, int count, String arg)
+  public void replaceData (int offset, int count, String arg)
     {
-      String data = getData();
-      setData(data.substring(0, offset) + arg + data.substring(offset +
-                                                               count));
+      String data = getData ();
+      setData (data.substring (0, offset) + arg +
+               data.substring (offset + count));
     }
 
 }
